+++ bespokeasm/assembler/model/__init__.py
@@ -13,6 +13,7 @@ from bespokeasm.assembler.keywords import ASSEMBLER_KEYWORD_SET
 from bespokeasm.assembler.model.instruction_set import InstructionSet
 from bespokeasm.assembler.model.operand_set import OperandSet, OperandSetCollection
 
+
 class AssemblerModel:
     _config: dict
 
@@ -34,11 +35,20 @@ class AssemblerModel:
         if 'min_version' in self._config['general']:
             required_version = self._config['general']['min_version']
             if is_verbose > 0:
-                click.echo(f'The ISA configuration file requires BespokeASM version {required_version}. This version of BespokeASM is {BESPOKEASM_VERSION_STR}.')
+                click.echo(
+                    f'The ISA configuration file requires BespokeASM version {required_version}. This '
+                    f'version of BespokeASM is {BESPOKEASM_VERSION_STR}.'
+                )
             if required_version > BESPOKEASM_VERSION_STR:
-                sys.exit(f'ERROR: the instruction set configuration file requires at least BespokeASM version {required_version}')
+                sys.exit(
+                    f'ERROR: the instruction set configuration file requires at least BespokeASM '
+                    f'version {required_version}'
+                )
             if required_version < BESPOKEASM_MIN_REQUIRED_STR:
-                sys.exit(f'ERROR: this version of BespokeASM requires a configuration file that minimally requires BespokeASM version {BESPOKEASM_MIN_REQUIRED_STR}')
+                sys.exit(
+                    f'ERROR: this version of BespokeASM requires a configuration file that minimally '
+                    f'requires BespokeASM version {BESPOKEASM_MIN_REQUIRED_STR}'
+                )
         # load ISA version information
         config_file_name = os.path.splitext(os.path.basename(config_file_path))[0]
         if 'identifier' in self._config['general']:
@@ -49,10 +59,13 @@ class AssemblerModel:
             version_match = re.match(
                 r"^\s*" + version.VERSION_PATTERN + r"\s*$",
                 self._isa_version,
-                flags=re.IGNORECASE|re.VERBOSE,
+                flags=re.IGNORECASE | re.VERBOSE,
             )
             if version_match is None:
-                sys.exit(f'ERROR - provide ISA version "{self._isa_version}" is not in semantic versioning format. See https://semver.org for details.')
+                sys.exit(
+                    f'ERROR - provide ISA version "{self._isa_version}" is not in semantic versioning '
+                    f'format. See https://semver.org for details.'
+                )
         else:
             self._isa_name = config_file_name
             self._isa_version = '0.0.1'
@@ -71,6 +84,7 @@ class AssemblerModel:
 
     def __repr__(self) -> str:
         return str(self)
+
     def __str__(self) -> str:
         if 'description' in self._config:
             desc_str = self._config['description']
@@ -81,26 +95,32 @@ class AssemblerModel:
     @property
     def description(self) -> str:
         return self._config.get('description', 'BespokeASM Assembly')
+
     @property
     def isa_name(self) -> str:
         '''Name of language defined by this ISA model. Defaults to configuration file basename.'''
         return self._isa_name
+
     @property
     def isa_version(self) -> str:
         '''Version number of this ISA model. Must be in semantic version format, e.g. "0.2.4"'''
         return self._isa_version
+
     @property
     def assembly_file_extenions(self) -> str:
         return self._file_extension
+
     @property
     def endian(self) -> str:
         if 'endian' in self._config['general']:
             return self._config['general']['endian']
         else:
             return 'big'
+
     @property
     def address_size(self) -> int:
         return self._config['general']['address_size']
+
     @property
     def registers(self) -> set[str]:
         return self._registers
@@ -108,6 +128,7 @@ class AssemblerModel:
     @property
     def instruction_mnemonics(self) -> list[str]:
         return list(self._instructions.keys())
+
     def get_operand_set(self, operand_set_name: str) -> OperandSet:
         return self._operand_sets.get_operand_set(operand_set_name)
 
@@ -116,7 +137,7 @@ class AssemblerModel:
         return self._config['general'].get('origin', 0)
 
     @property
-    def predefined_constants(self) -> list[dict[str,int]]:
+    def predefined_constants(self) -> list[dict[str, int]]:
         if 'predefined' in self._config and 'constants' in self._config['predefined']:
             return self._config['predefined']['constants']
         else: